Day 3 - The Basics of Querying the DOM
At this point, we know that we can search, or query the DOM using the CSS selectors that we’re already familiar with; but what about moving around the DOM after the fact? We’ll review a handful of traversal methods that you’ll use often in your projects, including children(), find(), parent(), closest(), and next().