Day 4: Events 101
Now, we get to move on to the fun part! In order to properly listen for when specific actions occur, such as the user clicking an anchor tag, or hovering over a particular section, we use event listeners. In this lesson, for demonstration purposes, we'll build a simple stylesheet switcher. Button troggle.